improve compatibility
